% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate clonotype abundance}
\usage{
calc_abundance(
  SCE_in,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  prefix = "",
  return_SCE = TRUE
)
}
\arguments{
\item{SCE_in}{SCE object containing V(D)J data}

\item{clonotype_col}{meta.data column containing clonotype IDs}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_SCE}{Return an SCE object. If set to FALSE, a tibble
summarizing the results is returned.}
}
\value{
Seurat object with clonotype abundance added to meta.data
}
\description{
Calculate clonotype abundance
}
