% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_similarity.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate repertoire similarity between clusters}
\usage{
calc_similarity(
  SCE_in,
  clonotype_col = "cdr3_nt",
  cluster_col,
  method = abdiv::jaccard,
  prefix = NULL,
  return_SCE = TRUE
)
}
\arguments{
\item{SCE_in}{Seurat object}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_SCE}{Return a Seurat object. If set to FALSE, a matrix is
returned}
}
\value{
Seurat object with similarity index added to meta.data
}
\description{
Calculate repertoire similarity between clusters
}
\author{
djvdj authors
}
