% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SCE.R
\name{.add_colData}
\alias{.add_colData}
\title{Add additional entries to the colData of an SCE}
\usage{
.add_colData(SCE.in, to_add, merge_on = NULL)
}
\arguments{
\item{SCE.in}{SCE object}

\item{to_add}{data.frame with additional colData entries}

\item{merge_on}{Indicate the column of \code{to_add}, which should contain the
entries corresponding to the colnames(SCE.in). Default: NULL, which will use
rownames of the \code{to_add} df.}
}
\value{
SCE object with amended colData entries
}
\description{
Add additional entries to the colData of an SCE
}
\author{
Friederike Dündar
}
