% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{import_vdj}
\alias{import_vdj}
\title{Import VDJ data to be added to an SCE object}
\usage{
import_vdj(
  vdj_dir,
  prefix = "",
  cell_prefix = "",
  filter_contigs = TRUE,
  sep = ";"
)
}
\arguments{
\item{vdj_dir}{cellranger VDJ output directories. If a vector of multiple
paths is provided, an equal number of cell prefixes must also be provided.
If a named vector is given, the names will be used to prefix each cell
barcode.}

\item{prefix}{Prefix to add to new \code{colData} columns}

\item{cell_prefix}{Prefix to add to cell barcodes}

\item{filter_contigs}{Only include chains with at least one productive
contig}

\item{sep}{Separator to use for storing per cell clonotype information in
the meta.data}

\item{return_SCE}{TRUE by default. If set to FALSE, only a data.frame will
be returned.}
}
\value{
Data.frame with the VDJ info that can be added to the colData of an SCE.
}
\description{
Import VDJ data to be added to an SCE object
}
