% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_SCE.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot repertoire overlap}
\usage{
plot_similarity(
  SCE_in,
  clonotype_col = "cdr3_nt",
  cluster_col,
  method = abdiv::jaccard,
  plot_colors = NULL,
  ...
)
}
\arguments{
\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap. Any grouping column from colData(SCE_in) will do.}

\item{method}{Method to use for calculating similarity between clusters}

\item{plot_colors}{Character vector containing colors for plotting}

\item{...}{Additional arguments to pass to ggplot}

\item{sce_in}{SCE object containing V(D)J data or matrix containing
similarity values to plot}
}
\value{
ggplot object
}
\description{
Plot repertoire overlap
}
\author{
djvdj authors
}
