% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvdj-package.R
\docType{package}
\name{SCEdjvdj-package}
\alias{SCEdjvdj}
\alias{SCEdjvdj-package}
\title{SCEdjvdj: Single-cell VDJ Tools for SCE objects}
\description{
djvdj provides a range of tools to analyze
    single-cell V(D)J data. Here the functions have been
    adapted to work with SingleCellExperiment objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/djvdj}
  \item \url{https://github.com/rnabioco/djvdj}
  \item Report bugs at \url{https://github.com/rnabioco/djvdj/issues}
}

}
\author{
\strong{Maintainer}: Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu} (\href{https://orcid.org/0000-0003-4012-3147}{ORCID})

Other contributors:
\itemize{
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} [contributor]
  \item Rui Fu \email{raysinensis@gmail.com} [contributor]
  \item Kent Riemondy \email{kent.riemondy@ucdenver.edu} [contributor]
}

}
\keyword{internal}
