% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_sce}
\alias{tiny_sce}
\title{SingleCellExperiment object for tests}
\format{
An SCE object with 100 features, 100 cells, and 1 assay
}
\usage{
tiny_sce
}
\description{
A tiny SCE object for tests
}
\keyword{datasets}
